.history/templates/fastapi-cloud/requirements-dev_20250906124011.txt
# Development Requirements for FastAPI Template
# Focused set for rapid hackathon development

# Core Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Security & Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.23
databases[postgresql]>=0.8.0
motor>=3.3.2
redis>=5.0.1

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Configuration
python-dotenv>=1.0.0
dynaconf>=3.2.4

# HTTP Client
httpx>=0.25.2

# Rate Limiting
slowapi>=0.1.9

# Logging
structlog>=23.2.0
loguru>=0.7.2

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Development Tools
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.1

# Utilities
click>=8.1.7
rich>=13.7.0
aiofiles>=23.2.1

# Cloud (basic)
boto3>=1.34.0

# Monitoring
prometheus-client>=0.19.0
psutil>=5.9.0
